[inductor] Add targeted unit tests for guards-related functions used in the codecache
Summary: Add a few unit tests that exercise `produce_guards_expression` and `evaluate_guards_expression` (and specifically "ToFloat" "FloatTrueDiv" added in https://github.com/pytorch/pytorch/pull/128418)

[ghstack-poisoned]
